transition to "naturally trimmed" name - old one is unfitting due to the new features
