+++ .github/ISSUE_TEMPLATE/incompatibility.yml
@@ -5,7 +5,7 @@ body:
   - type: textarea
     attributes:
       label: Incompatible Mod
-      description: List the mod that is incompatible with the Mob Armor Trims mod
+      description: List the mod that is incompatible with this mod
     validations:
       required: true
