+++ gradle.properties
@@ -3,11 +3,11 @@ org.gradle.jvmargs=-Xmx2G
 org.gradle.parallel=true
 
 # Mod properties
-mod.name=Mob Armor Trims
-mod.id=mob_armor_trims
+mod.name=Naturally Trimmed
+mod.id=naturally_trimmed
 mod.group=net.majo24
 mod.version=2.5.0
-mod.description=Makes mobs be able to spawn with naturally trimmed armor
+mod.description=Trims naturally generated armor and equipment from mobs, loot tables and trades
 mod.github_link=https://github.com/Imajo24I/Mob-Armor-Trims
 mod.issues_link=https://github.com/Imajo24I/Mob-Armor-Trims/issues/new/choose
