--- src/main/java/net/majo24/mob_armor_trims/config/Config.java
@@ -1,41 +1,0 @@
-package net.majo24.mob_armor_trims.config;
-
-import net.majo24.mob_armor_trims.MobArmorTrims;
-import net.majo24.mob_armor_trims.config.backend.ConfigManager;
-import net.majo24.mob_armor_trims.config.backend.annotations.Entry;
-import net.majo24.mob_armor_trims.config.backend.annotations.SubConfig;
-
-public class Config {
-    public static final ConfigManager<Config> CONFIG_MANAGER = new ConfigManager<>(Config.class, MobArmorTrims.getConfigPath());
-
-    @Entry(name = "enable_trim_mobs", comment = "Enables the equipment from mobs to be trimmed")
-    public boolean enableTrimMobs = true;
-
-    @Entry(name = "enable_trim_loot_tables", comment = "Enables the equipment generated by loot tables to be trimmed")
-    public boolean enableTrimLootTables = true;
-
-    @Entry(name = "enable_trim_trades", comment = "Enables the equipment from villager trades to be trimmed")
-    public boolean enableTrimTrades = true;
-
-    @SubConfig(name = "trim_mobs", comment = "Settings for trimming a mob's equipment")
-    public TrimMobsSubConfig trimMobs = new TrimMobsSubConfig();
-
-    @SubConfig(name = "trim_loot_tables", comment = "Settings for trimming equipment generated by loot tables")
-    public TrimLootTablesSubConfig trimLootTables = new TrimLootTablesSubConfig();
-
-    @SubConfig(name = "trim_trades", comment = "Settings for trimming equipment from villager trades")
-    public TrimTradesSubConfig trimTrades = new TrimTradesSubConfig();
-
-    public static class TrimLootTablesSubConfig {
-        @Entry(name = "trim_chance", comment = "Chance of an item generated by a loot table of having a trim applied")
-        public int trimChance = 50;
-    }
-
-    public static class TrimTradesSubConfig {
-        @Entry(name = "trim_chance", comment = "Chance of a villager trade of having a trim applied")
-        public int trimChance = 50;
-
-        @Entry(name = "min_level", comment = "The minimum villager trading level an item needs to be in, to have a chance of being trimmed")
-        public int minLevel = 3;
-    }
-}
