--- src/main/java/net/majo24/mob_armor_trims/config/backend/annotations/SubConfig.java
@@ -1,13 +1,0 @@
-package net.majo24.mob_armor_trims.config.backend.annotations;
-
-import java.lang.annotation.ElementType;
-import java.lang.annotation.Retention;
-import java.lang.annotation.RetentionPolicy;
-import java.lang.annotation.Target;
-
-@Retention(RetentionPolicy.RUNTIME)
-@Target({ElementType.FIELD})
-public @interface SubConfig {
-    String name();
-    String comment();
-}
