--- src/main/java/net/majo24/mob_armor_trims/config/screen/ModMenuIntegration.java
@@ -1,12 +1,0 @@
-package net.majo24.mob_armor_trims.config.screen;
-
-//? if fabric {
-import com.terraformersmc.modmenu.api.ConfigScreenFactory;
-import com.terraformersmc.modmenu.api.ModMenuApi;
-
-public class ModMenuIntegration implements ModMenuApi {
-    @Override
-    public ConfigScreenFactory<?> getModConfigScreenFactory() {
-        return ConfigScreenProvider::getConfigScreen;
-    }
-}//?}
