+++ src/main/java/net/majo24/naturally_trimmed/config/backend/annotations/Entry.java
@@ -1,0 +1,13 @@
+package net.majo24.naturally_trimmed.config.backend.annotations;
+
+import java.lang.annotation.ElementType;
+import java.lang.annotation.Retention;
+import java.lang.annotation.RetentionPolicy;
+import java.lang.annotation.Target;
+
+@Retention(RetentionPolicy.RUNTIME)
+@Target({ElementType.FIELD})
+public @interface Entry {
+    String name();
+    String comment() default "";
+}
