+++ src/main/java/net/majo24/naturally_trimmed/config/screen/ModMenuIntegration.java
@@ -1,0 +1,12 @@
+package net.majo24.naturally_trimmed.config.screen;
+
+//? if fabric {
+import com.terraformersmc.modmenu.api.ConfigScreenFactory;
+import com.terraformersmc.modmenu.api.ModMenuApi;
+
+public class ModMenuIntegration implements ModMenuApi {
+    @Override
+    public ConfigScreenFactory<?> getModConfigScreenFactory() {
+        return ConfigScreenProvider::getConfigScreen;
+    }
+}//?}
