--- src/main/resources/assets/mob_armor_trims/lang/en_us.json
@@ -1,58 +1,0 @@
-{
-  "mob_armor_trims.config.general": "General",
-  "mob_armor_trims.config.general.tooltip": "General settings for this mod",
-
-  "mob_armor_trims.config.general.enable_features": "Enable Features",
-  "mob_armor_trims.config.general.enable_features.description": "Settings for enabling specific features",
-  "mob_armor_trims.config.general.enable_features.enableTrimMobs": "Enable Mob Trimming",
-  "mob_armor_trims.config.general.enable_features.enableTrimMobs.description": "Enables the mob trimming feature",
-  "mob_armor_trims.config.general.enable_features.enableTrimLootTables": "Enable Loot Table Trimming",
-  "mob_armor_trims.config.general.enable_features.enableTrimLootTables.description": "Enables the Loot Table trimming feature",
-  "mob_armor_trims.config.general.enable_features.enableTrimTrades": "Enable Trade Trimming",
-  "mob_armor_trims.config.general.enable_features.enableTrimTrades.description": "Enables the Villager Trade Trimming feature",
-
-  "mob_armor_trims.config.general.trimLootTables": "Trim Loot Tables",
-  "mob_armor_trims.config.general.trimLootTables.description": "Settings for trimming items generated by loot tables",
-  "mob_armor_trims.config.general.trimLootTables.trimChance": "Trim Chance",
-  "mob_armor_trims.config.general.trimLootTables.trimChance.description": "Chance of an item generated by a loot table of having a trim applied",
-
-  "mob_armor_trims.config.general.trimTrades": "Trim Trades",
-  "mob_armor_trims.config.general.trimTrades.description": "Settings for trimming villager trades",
-  "mob_armor_trims.config.general.trimTrades.trimChance": "Trim Chance",
-  "mob_armor_trims.config.general.trimTrades.trimChance.description": "Chance of a trade of having a trim applied",
-  "mob_armor_trims.config.general.trimTrades.minLevel": "Min Level",
-  "mob_armor_trims.config.general.trimTrades.minLevel.description": "The minimum villager trading level a item needs to be in, to have a chance of being trimmed",
-
-
-  "mob_armor_trims.config.trimMobs": "Trim Mobs",
-  "mob_armor_trims.config.trimMobs.tooltip": "Settings for trimming a mobs equipment",
-  "mob_armor_trims.config.trimMobs.openCategory": "Open Trim Mob Category",
-  "mob_armor_trims.config.trimMobs.openCategory.description": "Opens the Trim Mob Category for further configuration",
-  "mob_armor_trims.config.trimMobs.trimSystem": "Trim System",
-  "mob_armor_trims.config.trimMobs.trimSystem.description": "Select the System of how to select, what trims to give mobs.\n- Random Trims: Randomly choose the trim, but also take the previous trim highly into account.\n- Custom Trims: Choose the trim out of a list of custom trims. You can manage the trims yourself",
-  "mob_armor_trims.config.trimMobs.noTrimsChance": "No Trims Chance",
-  "mob_armor_trims.config.trimMobs.noTrimsChance.description": "Chance of the mob having no trims at all",
-
-  "mob_armor_trims.config.trimMobs.randomTrims.description": "Settings for the Random Trims System.\nThese settings will only make a difference, if the random trims system is enabled.",
-  "mob_armor_trims.config.trimMobs.randomTrims.trimChance": "Trim Chance",
-  "mob_armor_trims.config.trimMobs.randomTrims.trimChance.description": "Chance of each armor piece of a mob having an armor trim",
-  "mob_armor_trims.config.trimMobs.randomTrims.similarTrimChance": "Similar Trim Chance",
-  "mob_armor_trims.config.trimMobs.randomTrims.similarTrimChance.description": "Chance of each armor piece having a similar armor trim as the previous armor piece",
-  "mob_armor_trims.config.trimMobs.randomTrims.blacklist": "Blacklist for Trim Patterns",
-  "mob_armor_trims.config.trimMobs.randomTrims.blacklist.description": "This is a blacklist for disabling trim patterns.\n\nRegex is used, which helps reducing a lot of repetition.\nThis also means, if you want to check for a specific pattern, you need to use \"^[pattern]$\" instead of just \"[pattern]\"\n\nBy default, all patterns from the Trimmable Tools mod are blacklisted, since they only work for tools and not armor",
-
-
-  "mob_armor_trims.config.utils": "Utils",
-  "mob_armor_trims.config.utils.tooltip": "Different utility buttons to make configuring the mod easier",
-  "mob_armor_trims.config.utils.reloadConfig": "Reload Config",
-  "mob_armor_trims.config.utils.reloadConfig.description": "Reload the config from the config file.\nUseful for configuring the Custom Trims Combination system, since it is only configurable through the config file",
-  "mob_armor_trims.config.utils.validateCustomTrimCombinations": "Validate Custom Trim Combinations",
-  "mob_armor_trims.config.utils.validateCustomTrimCombinations.description": "Validates the Custom Trim Combinations and logs each invalid combination in chat.\nOnly works while inside a world",
-  "mob_armor_trims.config.utils.validateBlacklist": "Validate Trim Blacklist",
-  "mob_armor_trims.config.utils.validateBlacklist.description": "Validates the Trim Blacklist and logs every blacklisted trim pattern, every non-blacklisted trim pattern and every unnecessary blacklist entry in chat.\nOnly works while inside a world",
-
-
-  "mob_armor_trims.config.backup_screen.installYACL": "Install Yet Another Config Lib to access the config screen",
-  "mob_armor_trims.config.backup_screen.viewOnModrinth": "View on Modrinth"
-
-}
