+++ src/main/resources/assets/naturally_trimmed/icon.png
