+++ src/main/resources/assets/naturally_trimmed/lang/en_us.json
@@ -1,0 +1,58 @@
+{
+  "naturally_trimmed.config.general": "General",
+  "naturally_trimmed.config.general.tooltip": "General settings for this mod",
+
+  "naturally_trimmed.config.general.enable_features": "Enable Features",
+  "naturally_trimmed.config.general.enable_features.description": "Settings for enabling specific features",
+  "naturally_trimmed.config.general.enable_features.enableTrimMobs": "Enable Mob Trimming",
+  "naturally_trimmed.config.general.enable_features.enableTrimMobs.description": "Enables the mob trimming feature",
+  "naturally_trimmed.config.general.enable_features.enableTrimLootTables": "Enable Loot Table Trimming",
+  "naturally_trimmed.config.general.enable_features.enableTrimLootTables.description": "Enables the Loot Table trimming feature",
+  "naturally_trimmed.config.general.enable_features.enableTrimTrades": "Enable Trade Trimming",
+  "naturally_trimmed.config.general.enable_features.enableTrimTrades.description": "Enables the Villager Trade Trimming feature",
+
+  "naturally_trimmed.config.general.trimLootTables": "Trim Loot Tables",
+  "naturally_trimmed.config.general.trimLootTables.description": "Settings for trimming items generated by loot tables",
+  "naturally_trimmed.config.general.trimLootTables.trimChance": "Trim Chance",
+  "naturally_trimmed.config.general.trimLootTables.trimChance.description": "Chance of an item generated by a loot table of having a trim applied",
+
+  "naturally_trimmed.config.general.trimTrades": "Trim Trades",
+  "naturally_trimmed.config.general.trimTrades.description": "Settings for trimming villager trades",
+  "naturally_trimmed.config.general.trimTrades.trimChance": "Trim Chance",
+  "naturally_trimmed.config.general.trimTrades.trimChance.description": "Chance of a trade of having a trim applied",
+  "naturally_trimmed.config.general.trimTrades.minLevel": "Min Level",
+  "naturally_trimmed.config.general.trimTrades.minLevel.description": "The minimum villager trading level a item needs to be in, to have a chance of being trimmed",
+
+
+  "naturally_trimmed.config.trimMobs": "Trim Mobs",
+  "naturally_trimmed.config.trimMobs.tooltip": "Settings for trimming a mobs equipment",
+  "naturally_trimmed.config.trimMobs.openCategory": "Open Trim Mob Category",
+  "naturally_trimmed.config.trimMobs.openCategory.description": "Opens the Trim Mob Category for further configuration",
+  "naturally_trimmed.config.trimMobs.trimSystem": "Trim System",
+  "naturally_trimmed.config.trimMobs.trimSystem.description": "Select the System of how to select, what trims to give mobs.\n- Random Trims: Randomly choose the trim, but also take the previous trim highly into account.\n- Custom Trims: Choose the trim out of a list of custom trims. You can manage the trims yourself",
+  "naturally_trimmed.config.trimMobs.noTrimsChance": "No Trims Chance",
+  "naturally_trimmed.config.trimMobs.noTrimsChance.description": "Chance of the mob having no trims at all",
+
+  "naturally_trimmed.config.trimMobs.randomTrims.description": "Settings for the Random Trims System.\nThese settings will only make a difference, if the random trims system is enabled.",
+  "naturally_trimmed.config.trimMobs.randomTrims.trimChance": "Trim Chance",
+  "naturally_trimmed.config.trimMobs.randomTrims.trimChance.description": "Chance of each armor piece of a mob having an armor trim",
+  "naturally_trimmed.config.trimMobs.randomTrims.similarTrimChance": "Similar Trim Chance",
+  "naturally_trimmed.config.trimMobs.randomTrims.similarTrimChance.description": "Chance of each armor piece having a similar armor trim as the previous armor piece",
+  "naturally_trimmed.config.trimMobs.randomTrims.blacklist": "Blacklist for Trim Patterns",
+  "naturally_trimmed.config.trimMobs.randomTrims.blacklist.description": "This is a blacklist for disabling trim patterns.\n\nRegex is used, which helps reducing a lot of repetition.\nThis also means, if you want to check for a specific pattern, you need to use \"^[pattern]$\" instead of just \"[pattern]\"\n\nBy default, all patterns from the Trimmable Tools mod are blacklisted, since they only work for tools and not armor",
+
+
+  "naturally_trimmed.config.utils": "Utils",
+  "naturally_trimmed.config.utils.tooltip": "Different utility buttons to make configuring the mod easier",
+  "naturally_trimmed.config.utils.reloadConfig": "Reload Config",
+  "naturally_trimmed.config.utils.reloadConfig.description": "Reload the config from the config file.\nUseful for configuring the Custom Trims Combination system, since it is only configurable through the config file",
+  "naturally_trimmed.config.utils.validateCustomTrimCombinations": "Validate Custom Trim Combinations",
+  "naturally_trimmed.config.utils.validateCustomTrimCombinations.description": "Validates the Custom Trim Combinations and logs each invalid combination in chat.\nOnly works while inside a world",
+  "naturally_trimmed.config.utils.validateBlacklist": "Validate Trim Blacklist",
+  "naturally_trimmed.config.utils.validateBlacklist.description": "Validates the Trim Blacklist and logs every blacklisted trim pattern, every non-blacklisted trim pattern and every unnecessary blacklist entry in chat.\nOnly works while inside a world",
+
+
+  "naturally_trimmed.config.backup_screen.installYACL": "Install Yet Another Config Lib to access the config screen",
+  "naturally_trimmed.config.backup_screen.viewOnModrinth": "View on Modrinth"
+
+}
