--- src/main/resources/mob_armor_trims.mixins.json
@@ -1,15 +1,0 @@
-{
-  "required": true,
-  "package": "net.majo24.mob_armor_trims.mixin",
-  "compatibilityLevel": "JAVA_17",
-  "mixins": [
-    "spawn_mobs_trimmed.SpawnMobsTrimmed",
-    "spawn_mobs_trimmed.SpawnSkeletonsTrimmed",
-    "spawn_mobs_trimmed.SpawnVindicatorsTrimmed",
-    "spawn_mobs_trimmed.SpawnZombiesTrimmed",
-    "VillagerTradesMixin"
-  ],
-  "injectors": {
-    "defaultRequire": 1
-  }
-}
