+++ src/main/resources/naturally_trimmed.accesswidener
