+++ src/main/resources/naturally_trimmed.mixins.json
@@ -1,0 +1,15 @@
+{
+  "required": true,
+  "package": "net.majo24.naturally_trimmed.mixin",
+  "compatibilityLevel": "JAVA_17",
+  "mixins": [
+    "spawn_mobs_trimmed.SpawnMobsTrimmed",
+    "spawn_mobs_trimmed.SpawnSkeletonsTrimmed",
+    "spawn_mobs_trimmed.SpawnVindicatorsTrimmed",
+    "spawn_mobs_trimmed.SpawnZombiesTrimmed",
+    "VillagerTradesMixin"
+  ],
+  "injectors": {
+    "defaultRequire": 1
+  }
+}
