+++ src/main/resources/pack.mcmeta
@@ -1,6 +1,6 @@
 {
   "pack": {
-    "description": "Mob Armor Trims",
+    "description": "Naturally Trimmed",
     "pack_format": 14
   }
 }
